[FEAT]: Customs를 한번에 받아서 저장하는 메서드 구현
