+++ pickme/src/main/java/project/pickme/user/repository/CustomsMapper.java
@@ -1,0 +1,17 @@
+package project.pickme.user.repository;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import project.pickme.user.domain.Customs;
+
+@Mapper
+public interface CustomsMapper {
+	void save(Customs customs);
+	List<Customs> findAll();
+	Optional<Customs> findById(String id);
+	void saveAll(List<Customs> customs);
+	void deleteAll();
+}
